I renamed the folder to reflect the changes
I changed the name of the task 2 folder to say "MedaillesOpdrachten", I also made a class for every upcoming task so I don't have to do that later. Finally, I made a start command so all I have to do is change the number after the word "task" when I want to run a specific task's code.