Task 1 transfer
I am pushing this separately from the other files because I finished one, and copied it over to this folder, basically needed a small explanation for that. To see the actual process for task 1, you'll have to check the task 1 folder. It should still exist for the sake of checking the initial development of that task.