A fix for one of the while loops
The first while loop that runs the whole game was wrong. It has to end the code when one of the two (Win/Lose) conditions becomes true. I accidentally put an && instead, making it so that the game only ends when both are true. Now you can see on line 37 I added || instead.