Klas Medals
These aren't actual tasks related to the ones I have to turn in. These are tasks given during class that are saved here for the sake of logging everything I do in the class. It's easier to simply add them to the relevant term's set of tasks than to make a unique repository for it, so here it is.