General changes and updates
I streamlined the array in task 7, and I changed task 6 to have a "universal minimum" which is just 1, and made it so the user cannot guess under 1 with it still counting. The other change is that I renamed one of the class tasks so it wouldn't get in the way, and I turned them all into notes so that I can just redo them inside of the tasks, instead of pulling from these class examples.