Task notes and organization
Noted down which tasks are complete, turned in, haven't been started yet, and which ones are incomplete. Also added notes to say what I have to do for the incomplete ones. Soon I'll do the same for the ones that haven't been started so I don't have to keep going back to Canvas to check.