Major changes to the main code
This contains a lot of major changes for the sake of convenience and preservation.

I made a class called PorgramLeftovers in order to have a place to easily save and access all of the code made during class that was cluttering up the main Program Class.

Secondly, I finished task 8, making a constructor within the task, and writing the code within the Program to make the characters through user input.

Finally, I changed the program to a switch case within a while loop, where you can select which program you want to start by inputting a number through the console. This makes it easier to start up a new task without me having to remove code, or constantly change a "task.Start" every time I want to work on or show a new task.

Currently there are no errors and it all functions well. I hope that it continues this way, because this is a nice solution to a small problem that made it annoying to switch between tasks to check.

To summarize, I finished task 8 and moved around a bunch of code to make starting it, and other tasks more convenient.