Formatting & Task 9
I just made some changes to organize the code, separating variables from specific tasks that take up the Program Class, and making it clear which variables are shared among these tasks. This is also the first version of task 9. Not sure if it's finished yet, as I worked on it yesterday as of writing this.