Name change and efficiency changes Task 8
I changed the name of task 8 to Character, as requested in the task. I also changed the code so that each class is added into a list.
Finally, I change it so that they get printed into a foreach loop, so the ShowStats command only has to be typed in once.